DTSCCI-2278: Async processing of case references
