+++ src/test/java/uk/gov/hmcts/reform/civil/handler/migration/MigrateCasesEventHandlerTest.java
@@ -10,26 +10,17 @@ import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 import uk.gov.hmcts.reform.civil.bulkupdate.csv.CaseReference;
 import uk.gov.hmcts.reform.civil.bulkupdate.csv.CaseReferenceCsvLoader;
-import uk.gov.hmcts.reform.civil.callback.CaseEvent;
-import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.ExternalTaskData;
-import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 import uk.gov.hmcts.reform.civil.utils.CaseMigrationEncryptionUtil;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.anyString;
-import static org.mockito.Mockito.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -41,30 +32,22 @@ class MigrateCasesEventHandlerTest {
     @Mock
     private CaseReferenceCsvLoader caseReferenceCsvLoader;
 
-    @Mock
-    private CoreCaseDataService coreCaseDataService;
-
-    @Mock
-    private CaseDetailsConverter caseDetailsConverter;
     private ObjectMapper objectMapper = new ObjectMapper();
 
     @Mock
     private MigrationTaskFactory migrationTaskFactory;
+    @Mock
+    private AsyncCaseMigrationService asyncCaseMigrationService;
 
     private MigrateCasesEventHandler handler;
 
     @BeforeEach
     void setUp() {
-        caseDetailsConverter = new CaseDetailsConverter(objectMapper);
         MockitoAnnotations.openMocks(this);
         handler = new MigrateCasesEventHandler(
             caseReferenceCsvLoader,
-            coreCaseDataService,
-            caseDetailsConverter,
             migrationTaskFactory,
-            objectMapper,
-            500, // migrationBatchSize
-            10, // migrationWaitTime
+            asyncCaseMigrationService,
             "DUMMY_KEY"
         );
     }
@@ -80,31 +63,19 @@ class MigrateCasesEventHandlerTest {
         @SuppressWarnings("unchecked")
         MigrationTask<CaseReference> migrationTask = mock(MigrationTask.class);
         when(migrationTask.getType()).thenReturn(CaseReference.class);
-        when(migrationTask.getEventDescription()).thenReturn("Test Migration Task");
-        when(migrationTask.getEventSummary()).thenReturn("Migrating cases for test task");
 
         when(migrationTaskFactory.getMigrationTask("testTask")).thenReturn(Optional.of(migrationTask));
-
-        StartEventResponse startEventResponse = mock(StartEventResponse.class);
-        when(coreCaseDataService.startUpdate(anyString(), eq(CaseEvent.UPDATE_CASE_DATA))).thenReturn(startEventResponse);
-
-        uk.gov.hmcts.reform.ccd.client.model.CaseDetails caseDetails = mock(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.class);
-        when(startEventResponse.getCaseDetails()).thenReturn(caseDetails);
 
         CaseData caseData = mock(CaseData.class);
 
         List<CaseReference> mockReferences = List.of(new CaseReference("12345"), new CaseReference("67890"));
         when(caseReferenceCsvLoader.loadCaseReferenceList(CaseReference.class, csvFileName)).thenReturn(mockReferences);
 
-        when(caseDetailsConverter.toCaseData(startEventResponse.getCaseDetails())).thenReturn(caseData);
-        when(migrationTask.migrateCaseData(any(CaseData.class), any(CaseReference.class))).thenReturn(caseData);
-
         ExternalTaskData result = handler.handleTask(externalTask);
 
         // Assert
         assertNotNull(result);
-        verify(coreCaseDataService, times(2)).startUpdate(anyString(), eq(CaseEvent.UPDATE_CASE_DATA));
-        verify(coreCaseDataService, times(2)).submitUpdate(anyString(), any(CaseDataContent.class));
+        verify(asyncCaseMigrationService, times(1)).migrateCasesAsync(migrationTask, mockReferences);
     }
 
     @Test
@@ -150,7 +121,6 @@ class MigrateCasesEventHandlerTest {
     @Test
     void shouldReturnEmptyListWhenCaseIdsAndCsvFileNameAreNull() {
         // Arrange
-        String caseIds = null;
         String csvFileName = null;
 
         // Act
@@ -160,30 +130,4 @@ class MigrateCasesEventHandlerTest {
         assertEquals(0, result.size());
     }
 
-    @Test
-    void shouldBuildCaseDataContentSuccessfully() {
-        // Arrange
-        StartEventResponse startEventResponse = mock(StartEventResponse.class);
-        when(startEventResponse.getToken()).thenReturn("testToken");
-        when(startEventResponse.getEventId()).thenReturn("testEventId");
-
-        CaseData caseData = mock(CaseData.class);
-        Map<String, Object> mockData = Map.of("key1", "value1", "key2", "value2");
-        when(caseData.toMap(any())).thenReturn(mockData);
-
-        MigrationTask migrationTask = mock(MigrationTask.class);
-        when(migrationTask.getEventDescription()).thenReturn("Test Migration Task");
-        when(migrationTask.getEventSummary()).thenReturn("Migrating cases for test task");
-
-        // Act
-        CaseDataContent result = handler.buildCaseDataContent(startEventResponse, caseData, migrationTask);
-
-        // Assert
-        assertEquals("testToken", result.getEventToken());
-        assertEquals("testEventId", result.getEvent().getId());
-        assertEquals(mockData, result.getData());
-
-        assertEquals("Test Migration Task", result.getEvent().getDescription());
-        assertEquals("Migrating cases for test task", result.getEvent().getSummary());
-    }
 }
